[HAUSSDORFF] Refactor to support Numpy and OpenCV formats + add datatype parameter.

+++ common/contour_utils.py
@@ -11,10 +11,10 @@ def get_contours(bin_image, method=cv2.CHAIN_APPROX_NONE):
     return contours, hierarchy
 
 
-def sample_points_from_contour(contours, opencv_fmt=False):
+def sample_points_from_contour(contours, opencv_fmt=False, dtype=np.uint16):
     # Assumes that the input is a simplified contour.
     if opencv_fmt:
-        points = np.unique(np.vstack(contours).astype(np.uint16), axis=0)
+        points = np.unique(np.vstack(contours), axis=0)
     else:
-        points = swap_cols(np.unique(np.vstack(contours).astype(np.uint16).reshape([-1, 2]), axis=0))
-    return points
+        points = swap_cols(np.unique(np.vstack(contours).reshape([-1, 2]), axis=0))
+    return points.astype(dtype)
